+++ cmd/server/main.go
@@ -8,7 +8,7 @@ import (
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/chi/v5/middleware"
 
-	"reby/app"
+	backend "github.com/HectorMRC/backend-tech-test"
 )
 
 const httpPort = 8080
@@ -17,8 +17,8 @@ func main() {
 	r := chi.NewRouter()
 	r.Use(middleware.Logger)
 
-	r.Post("/rides", app.RideStartHandler)
-	r.Post("/rides/{id}/finish", app.RideFinishHandler)
+	r.Post("/rides", backend.RideStartHandler)
+	r.Post("/rides/{id}/finish", backend.RideFinishHandler)
 
 	if err := http.ListenAndServe(fmt.Sprintf(":%d", httpPort), r); err != http.ErrServerClosed && err != nil {
 		log.Fatalf("Error starting http server <%s>", err)
