+++ go.mod
@@ -1,5 +1,3 @@
-module reby
+module github.com/HectorMRC/backend-tech-test
 
-go 1.16
-
-require github.com/go-chi/chi/v5 v5.0.1
+go 1.18
